+++ src/main/java/com/itstime/xpact/domain/experience/dto/request/ExperienceUpdateRequestDto.java
@@ -1,34 +1,32 @@
-package com.itstime.xpact.domain.experience.dto;
+package com.itstime.xpact.domain.experience.dto.request;
 
 import com.itstime.xpact.domain.experience.common.ExperienceType;
 import com.itstime.xpact.domain.experience.common.FormType;
 import com.itstime.xpact.domain.experience.common.Status;
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 import java.time.LocalDate;
 
 @Getter
-@AllArgsConstructor
-@Schema(description = "경험 생성 요청 DTO")
-public class ExperienceCreateRequestDto {
+@Schema(description = "경험 수정 요청 DTO")
+public class ExperienceUpdateRequestDto {
 
     // common (공통 부분)
-    @Schema(description = "경험 저장 방식 지정 (임시저장 or 저장)", example = "SAVE", allowableValues = {"SAVE", "DRAFT"})
+    @Schema(description = "경험 저장 방식 지정 (임시저장 or 저장)", example = "SAVE", allowableValues = {"STASH", "SAVE"})
     private Status status;
 
     @Schema(description = "경험 유형 지정",
-            example = "INTERN",
+            example = "PROJECT",
             allowableValues = {"INTERN", "EXTERNAL_ACTIVITIES", "CONTEST",
                     "PROJECT", "CERTIFICATES", "ACADEMIC_CLUB", "EDUCATION",
-                    "PRIZE", "VOLUNTEER_WORK", "STUDY_ABROAD", "ETC"})
+                    "VOLUNTEER_WORK", "STUDY_ABROAD", "ETC"})
     private ExperienceType experienceType;
 
-    @Schema(description = "경험 양식 지정 (Star양식 or 간결 양식)", example = "STAR_FORM")
+    @Schema(description = "경험 양식 지정 (Star양식 or 간결 양식", example = "SIMPLE_FORM")
     private FormType formType;
 
-    @Schema(description = "경험 제목", example = "제목을 입력하세요")
+    @Schema(description = "경험 제목", example = "경험 제목")
     private String title;
 
     @Schema(description = "시작 일시", example = "2025-03-27")
