+++ src/main/java/com/itstime/xpact/domain/experience/dto/response/DetailExperienceReadResponseDto.java
@@ -1,4 +1,4 @@
-package com.itstime.xpact.domain.experience.dto;
+package com.itstime.xpact.domain.experience.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.itstime.xpact.domain.experience.common.ExperienceType;
